+++ generators/model/form.php
@@ -1,4 +1,7 @@
 <?php
+
+use yii\gii\generators\model\Generator;
+
 /* @var $this yii\web\View */
 /* @var $form yii\widgets\ActiveForm */
 /* @var $generator yii\gii\generators\form\Generator */
@@ -9,7 +12,11 @@ echo $form->field($generator, 'ns');
 echo $form->field($generator, 'baseClass');
 echo $form->field($generator, 'db');
 echo $form->field($generator, 'useTablePrefix')->checkbox();
-echo $form->field($generator, 'generateRelations')->checkbox();
+echo $form->field($generator, 'generateRelations')->dropDownList([
+    Generator::RELATIONS_NONE => Yii::t('yii', 'No relations'),
+    Generator::RELATIONS_ALL => Yii::t('yii', 'All relations'),
+    Generator::RELATIONS_ALL_INVERSE => Yii::t('yii', 'All relations with inverse'),
+]);
 echo $form->field($generator, 'generateLabelsFromComments')->checkbox();
 echo $form->field($generator, 'generateQuery')->checkbox();
 echo $form->field($generator, 'queryNs');